Added data caching and experimental delta values + other maintenance
Data caching done by storing the individually read bytes into a
list. This enables calculating deltas of various data values to make
it possible to understand the control logic further.

The match-case mess in function printByte was cleaned up into order of
the data bytes as they appear in the datastream documentation. This
hopefully makes it easier to find the relevant spots in the source
code for further development.