Bugfix - program could get stuck in sync phase
In some rare cases if the data stream ends during the sync phase in
such a manner that resync was started, any keyboard input was being
ignored and the program would appear stuck. In those cases the only
way to terminate the program was ^C.

The reason was no keyboard input processing within the sync loop.

Commonly this could also happen immediately after starting the program
if the data fed to it contained nothing to sync to.

In the rare cases after losing sync during usual operation this was
evident as the packet position ticker next to the sync state would be
stopped over the sync phase (over the seven 's's).