New feature: Basic seeking functions with hjkl implemented
When using a data capture file as the data source, seeking can be
performed with the following keys:

    h	  seek backwards 60 packets
    j	  seek backwards 10 packets
    k	  seek forwards 10 packets
    l	  seek forwards 60 packets

10 packets is roughly 15 seconds in real time. 60 packets is about 90
seconds, or a minute and a half. The seeking is done in multiples of
packet sizes to avoid unnecessary resyncing. However, the data on
screen will not be fully valid until a whole packet is read to update
all of the data shown on screen.

A position counter is shown next to the packet ticker to aid in
seeking to a specific part of a data capture. Seeking will stop
functioning once EOF is reached.

The counter and seeking functions are not available when the decoder
decodes live serial data.